gegl-parallel: fix thread-time comparison
s/gdouble/gint64/ in the thread-time comparison function.
Incidentally, it shouldn't actually affect the result.

+++ gegl/gegl-parallel.c
@@ -499,8 +499,8 @@ static gint
 gegl_parallel_distribute_update_thread_time_compare (gconstpointer x,
                                                      gconstpointer y)
 {
-  return (*(const gdouble *) x > *(const gdouble *) y) -
-         (*(const gdouble *) x < *(const gdouble *) y);
+  return (*(const gint64 *) x > *(const gint64 *) y) -
+         (*(const gint64 *) x < *(const gint64 *) y);
 }
 
 static void
